requirements.txt
selenium
beautifulsoup4
Pillow